Use get targets from name function consistently
